# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MentorNepal API",
    "version": "1.0.0",
    "description": "API docs for MentorNepal backend"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/onboarding/role": {
      "post": {
        "summary": "Select user role",
        "tags": [
          "Onboarding"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "role"
                ],
                "properties": {
                  "role": {
                    "type": "string",
                    "enum": [
                      "mentor",
                      "mentee"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Role selected successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "role": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/onboarding/mentor": {
      "post": {
        "summary": "Create or update mentor profile",
        "tags": [
          "Onboarding"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "fullName",
                  "headline",
                  "bio",
                  "expertise",
                  "yearsExp"
                ],
                "properties": {
                  "fullName": {
                    "type": "string"
                  },
                  "headline": {
                    "type": "string"
                  },
                  "bio": {
                    "type": "string"
                  },
                  "expertise": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "yearsExp": {
                    "type": "integer"
                  },
                  "hourlyRate": {
                    "type": "integer"
                  },
                  "company": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  },
                  "avatarUrl": {
                    "type": "string",
                    "format": "uri"
                  },
                  "linkedinUrl": {
                    "type": "string",
                    "format": "uri"
                  },
                  "location": {
                    "type": "string"
                  },
                  "languages": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mentor profile created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/onboarding/mentee": {
      "post": {
        "summary": "Create or update mentee profile",
        "tags": [
          "Onboarding"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "fullName",
                  "goals",
                  "careerStage"
                ],
                "properties": {
                  "fullName": {
                    "type": "string"
                  },
                  "goals": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "careerStage": {
                    "type": "string",
                    "enum": [
                      "student",
                      "early",
                      "mid",
                      "senior"
                    ]
                  },
                  "interests": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "bio": {
                    "type": "string"
                  },
                  "avatarUrl": {
                    "type": "string",
                    "format": "uri"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mentee profile created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/onboarding/complete": {
      "post": {
        "summary": "Complete onboarding",
        "tags": [
          "Onboarding"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Onboarding completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
...
}